requirements.txt
python-docx==0.8.11
streamlit
python-docx-replace
requests==2.28.1
mysql-connector-python==8.0.31
